Move `DataflowNodeToSqlSubqueryVisitor` to a separate file.
