+++ metricflow/dataflow/builder/node_data_set.py
@@ -12,9 +12,7 @@ from metricflow.dataflow.dataflow_plan import (
     DataflowPlanNode,
 )
 from metricflow.dataset.sql_dataset import SqlDataSet
-from metricflow.plan_conversion.to_sql_plan.dataflow_to_sql import (
-    DataflowNodeToSqlSubqueryVisitor,
-)
+from metricflow.plan_conversion.to_sql_plan.dataflow_to_subquery import DataflowNodeToSqlSubqueryVisitor
 
 if TYPE_CHECKING:
     from metricflow_semantics.model.semantic_manifest_lookup import SemanticManifestLookup
